+++ arch/x86_64/entry.cpp
@@ -301,7 +301,8 @@ boot::BootInfo build_boot_info() {
 
 } // namespace
 
-extern "C" [[noreturn]] void _start() {
+extern "C" [[noreturn]] __attribute__((no_stack_protector)) void _start() {
+    arch::x86_64::initialize_stack_chk_guard();
     console::early_init();
 
     if (!LIMINE_BASE_REVISION_SUPPORTED(g_limine_base_revision)) {